+++ code/test/sixsq/nuvla/server/resources/nuvlabox_status_2_lifecycle_test.clj
@@ -339,7 +339,7 @@
                                      (ltu/body)
                                      :next-heartbeat))]
 
-          (testing "online flag is propagated to nuvlabox"
+          (testing "online flag is denormlized to nuvlabox"
             (-> session-admin
                 (request nuvlabox-url)
                 (ltu/body->edn)
@@ -363,7 +363,7 @@
           is not visible from cimi (blacklisted)"
             (is (= true (:online-prev (db/retrieve state-id {})))))
 
-          (testing "online flag propagated to nuvlabox"
+          (testing "online flag denormalized to nuvlabox"
             (-> session-admin
                 (request nuvlabox-url)
                 (ltu/body->edn)